fix(insights-ui): read home-page top industries from DB, not HTTP self-fetch
Root cause of the failed Vercel deploy: during `next build` the home page
("/") self-fetched its own API at
`getBaseUrlForServerSidePages()/api/.../only-industries`. On Vercel
`getBaseUrl()` is empty server-side, so that resolves to the public canonical
origin https://koalagains.com. Post-migration that origin is fronted by
CloudFront→AWS, so the build reached across the CDN to a different production
deployment, which returned an HTML error/redirect page instead of JSON. The
home page's `res.json()` then threw "Unexpected token '<'" during static
export and aborted the whole build (the same HTML-instead-of-JSON also hit the
college-ambassador and tariff-listing fetches in the log).

Fix the root cause instead of masking it: the endpoint is a pure first-party
Prisma read, so extract its query into a shared server-only helper
getTopIndustriesWithTickers() and call it directly from the Server Component.
No base URL, no CDN, no cross-deployment dependency, always fresh on every
platform — and the showcase is only ever empty when the DB truly has no
industries. The /api route is kept as a thin wrapper over the same helper with
byte-identical behavior (it still hardcodes US exchanges). Tag-based
revalidation is preserved via the existing unstable_cache tag.
