+++ insights-ui/src/app/page.tsx
@@ -5,12 +5,12 @@ import { Hero } from '@/components/home-page/Hero';
 import KoalagainsOfferings from '@/components/home-page/KoalagainsOfferings';
 import KoalaGainsPlatform from '@/components/home-page/KoalaGainsPlatform';
 import ServiceNavigation from '@/components/home-page/ServiceNavigation';
-import { IndustryWithTopTickers, OnlyIndustriesResponse } from '@/types/api/ticker-industries';
+import { IndustryWithTopTickers } from '@/types/api/ticker-industries';
 import { KoalaGainsSpaceId } from '@/types/koalaGainsConstants';
 import { getPostsData } from '@/util/blog-utils';
 import { themeColors } from '@/util/theme-colors';
 import { SupportedCountries } from '@/utils/countryExchangeUtils';
-import { getBaseUrlForServerSidePages } from '@/utils/getBaseUrlForServerSidePages';
+import { getTopIndustriesWithTickers } from '@/utils/home-page/top-industries';
 import { getStocksPageTag, getHomePagePostsTag } from '@/utils/ticker-v1-cache-utils';
 import type { Metadata } from 'next';
 import { unstable_cache } from 'next/cache';
@@ -123,35 +123,12 @@ export const metadata: Metadata = {
 const WEEK = 60 * 60 * 24 * 7;
 
 async function fetchTopIndustriesWithTickers(): Promise<IndustryWithTopTickers[]> {
-  const baseUrl = getBaseUrlForServerSidePages();
-
-  const url = `${baseUrl}/api/${KoalaGainsSpaceId}/tickers-v1/country/${SupportedCountries.US}/tickers/only-industries`;
-
-  // Also tag the underlying fetch so any manual tag revalidation hits this too
-  try {
-    const res = await fetch(url, { next: { tags: [getStocksPageTag(SupportedCountries.US)] } });
-    if (!res.ok) return [];
-
-    // On Vercel the build self-fetches the production origin (getBaseUrlForServerSidePages()
-    // falls back to https://koalagains.com). If that origin momentarily serves an HTML
-    // error/redirect page instead of JSON, `res.json()` throws "Unexpected token '<'" and,
-    // because this runs during the static export of "/", aborts the entire build. Guard on the
-    // content-type and degrade to an empty list — the same intent as the `!res.ok` branch above.
-    const contentType = res.headers.get('content-type') ?? '';
-    if (!contentType.includes('application/json')) {
-      console.error(`Expected JSON from ${url} but received content-type "${contentType}" — rendering home page without top industries`);
-      return [];
-    }
-
-    const resp: OnlyIndustriesResponse = await res.json();
-
-    return resp.industries;
-  } catch (e) {
-    // Transient network/parse failures must not fail the build: render the home page without
-    // the top-industries section and let the next revalidation backfill it.
-    console.error('Failed to fetch top industries for home page:', e);
-    return [];
-  }
+  // Read the DB directly (server-only) instead of self-fetching our own /api over HTTP. During
+  // `next build` that self-fetch resolved to the public canonical origin (https://koalagains.com,
+  // now CloudFront→AWS), which could return an HTML error page and crash `res.json()`, aborting the
+  // static export of "/". A direct query always returns fresh data on every platform.
+  const { industries } = await getTopIndustriesWithTickers(KoalaGainsSpaceId, SupportedCountries.US);
+  return industries;
 }
 
 // Cache + tag BOTH data sources for 7 days
